refactor: move map components up
